Fix - Rename components, fix grid item minimum width size
