+++ src/pages/CreateTrackOfBest/ContentTemplete.js
@@ -1,20 +1,13 @@
 import ContentEditor from "components/ContentEditor"
 import InfoTooltip from "components/InfoTooltip"
 import { OutlineInput } from "components/Inputs"
-import TagSelectInput from "components/TagSelectInput"
-import { isEmpty } from "lodash-es"
 import React from "react"
 import styled from "styled-components"
-import {
-  borderColor,
-  secondaryTextColor,
-  warningColor,
-  backgroundColor1,
-  interactionColor,
-} from "styles/colors"
-import { title1Normal, body3Normal, body2Bold, body3Bold } from "styles/textTheme"
-import CoverImageContainer from "./CoverImageContainer"
-import YoutubeVideoContent from "./YoutubeVideoContainer"
+import { borderColor, secondaryTextColor, warningColor } from "styles/colors"
+import { title1Normal, body3Normal } from "styles/textTheme"
+import CoverImageContainer from "./CoverImageSectionContainer"
+import TagSectionContainer from "./TagSectionContainer"
+import YoutubeVideoContent from "./YoutubeVideoSectionContainer"
 
 const Container = styled.div`
   display: flex;
@@ -46,7 +39,7 @@ const SectionContainer = styled.div`
   }
 `
 
-const RowContainer = styled.div`
+const InputRowContainer = styled.div`
   display: flex;
   width: 100%;
   & > :not(:last-child) {
@@ -59,56 +52,29 @@ const RowContainer = styled.div`
   }
 `
 
-const PopularTagContainer = styled.div`
-  display: flex;
-  flex-flow: column;
-  width: 200px;
-  min-height: 200px;
-  border: 1px dashed ${secondaryTextColor};
-  background: ${backgroundColor1};
-  padding: 18px 16px 8px 0px;
-  margin-left: 40px;
-  .title {
-    padding: 0px 16px 16px;
-    ${body2Bold}
-  }
-  .tagContainer {
-    display: flex;
-    width: 100%;
-    height: 100%;
-    & > div {
-      margin-left: 16px;
-      margin-bottom: 10px;
-    }
-    flex-wrap: wrap;
-  }
-  .emptyView {
-    display: flex;
-    width: 100%;
-    height: 100%;
-    justify-content: center;
-    align-items: center;
-    ${body3Normal}
-    color: ${secondaryTextColor};
-  }
-`
+/**
+ * ContentTemplete
+ * @param {ReactNode} headerSection
+ * @param {Observable<List<Object>>} form
+ * @param {Function} onChange
+ * @param {Function} setError
+ * @param {Observable<File>} imageFile
+ * @param {Function} setImageFile
+ * @param {List<String>} tagOptions
+ * @param {List<String>} popularTagOptions
+ * @param {Object} defaultValue
+ * @param {String} mode "create" or "update"
+ */
 
-const TagItem = styled.div`
-  display: flex;
-  padding: 6px 12px;
-  ${body3Bold}
-  background: ${interactionColor};
-`
-
-export default function ContentContainer({
+export default function ContentTemplete({
   headerSection,
-  onChange,
   form,
-  setImageFile,
+  onChange,
+  setError,
   imageFile,
+  setImageFile,
   tagOptions = [],
   popularTagOptions = [],
-  setError,
   defaultValue,
   mode,
 }) {
@@ -117,7 +83,7 @@ export default function ContentContainer({
       title: "가수명 / 곡제목",
       contentBuilder: () => {
         return (
-          <RowContainer>
+          <InputRowContainer>
             <OutlineInput
               className="input"
               labelText="가수명"
@@ -134,7 +100,7 @@ export default function ContentContainer({
               onChange={(value) => onChange("songName", value)}
               defaultValue={defaultValue?.songName || form.fields.songName.value}
             />
-          </RowContainer>
+          </InputRowContainer>
         )
       },
       required: true,
@@ -217,31 +183,12 @@ export default function ContentContainer({
       title: "태그 선택",
       contentBuilder: () => {
         return (
-          <div style={{ display: "flex", width: "100%", paddingBottom: "20px" }}>
-            <TagSelectInput
-              options={tagOptions.map((item) => {
-                return { value: item }
-              })}
-              placeholder="태그를 선택해주세요"
-              notFoundContent={<div>존재하지 않는 태그명입니다.</div>}
-              width={300}
-              onChange={(value) => onChange("tag", value)}
-              value={defaultValue?.tag || form.fields.tag.value || []}
-              mode="tags"
-            />
-            <PopularTagContainer>
-              <div className="title">🔥 인기태그 🔥</div>
-              <div className="tagContainer">
-                {isEmpty(popularTagOptions) ? (
-                  <div className="emptyView">인기 태그가 없습니다.</div>
-                ) : (
-                  popularTagOptions.map((item) => {
-                    return <TagItem>{item}</TagItem>
-                  })
-                )}
-              </div>
-            </PopularTagContainer>
-          </div>
+          <TagSectionContainer
+            tagOptions={tagOptions}
+            popularTagOptions={popularTagOptions}
+            onChange={(value) => onChange("tag", value)}
+            defaultValue={defaultValue?.tag || form.fields.tag.value || []}
+          />
         )
       },
       infoText: (
