+++ src/pages/CreateTrackOfBest/CoverImageSectionContainer.js
