+++ src/utils/imageCropper.js
@@ -17,6 +17,7 @@ function getRadianAngle(degreeValue) {
  * @param {Object} pixelCrop - pixelCrop Object provided by react-easy-crop
  * @param {number} rotation - optional rotation parameter
  */
+
 export default async function getCroppedImg(imageSrc, pixelCrop, rotation = 0) {
   const image = await createImage(imageSrc)
   const canvas = document.createElement("canvas")
@@ -53,9 +54,6 @@ export default async function getCroppedImg(imageSrc, pixelCrop, rotation = 0) {
     Math.round(0 - safeArea / 2 + image.width * 0.5 - pixelCrop.x),
     Math.round(0 - safeArea / 2 + image.height * 0.5 - pixelCrop.y)
   )
-
-  // As Base64 string
-  // return canvas.toDataURL('image/jpeg');
 
   // As a blob
   return new Promise((resolve) => {
